+++ editor/model/actions/MoveGraphPointAction.cpp
@@ -6,14 +6,14 @@
 
 using namespace AxiomModel;
 
-MoveGraphPointAction::MoveGraphPointAction(const QUuid &controlUuid, uint8_t index, float oldTime, float oldValue,
-                                           float newTime, float newValue, AxiomModel::ModelRoot *root)
+MoveGraphPointAction::MoveGraphPointAction(const QUuid &controlUuid, uint8_t index, double oldTime, double oldValue,
+                                           double newTime, double newValue, AxiomModel::ModelRoot *root)
     : Action(ActionType::MOVE_GRAPH_POINT, root), _controlUuid(controlUuid), _index(index), _oldTime(oldTime),
       _oldValue(oldValue), _newTime(newTime), _newValue(newValue) {}
 
 std::unique_ptr<MoveGraphPointAction> MoveGraphPointAction::create(const QUuid &controlUuid, uint8_t index,
-                                                                   float oldTime, float oldValue, float newTime,
-                                                                   float newValue, AxiomModel::ModelRoot *root) {
+                                                                   double oldTime, double oldValue, double newTime,
+                                                                   double newValue, AxiomModel::ModelRoot *root) {
     return std::make_unique<MoveGraphPointAction>(controlUuid, index, oldTime, oldValue, newTime, newValue, root);
 }
