+++ editor/model/actions/SetNumRangeAction.cpp
@@ -6,13 +6,13 @@
 
 using namespace AxiomModel;
 
-SetNumRangeAction::SetNumRangeAction(const QUuid &uuid, float beforeMin, float beforeMax, uint32_t beforeStep,
-                                     float afterMin, float afterMax, uint32_t afterStep, AxiomModel::ModelRoot *root)
+SetNumRangeAction::SetNumRangeAction(const QUuid &uuid, double beforeMin, double beforeMax, uint32_t beforeStep,
+                                     double afterMin, double afterMax, uint32_t afterStep, AxiomModel::ModelRoot *root)
     : Action(ActionType::SET_NUM_RANGE, root), _uuid(uuid), _beforeMin(beforeMin), _beforeMax(beforeMax),
       _beforeStep(beforeStep), _afterMin(afterMin), _afterMax(afterMax), _afterStep(afterStep) {}
 
-std::unique_ptr<SetNumRangeAction> SetNumRangeAction::create(const QUuid &uuid, float beforeMin, float beforeMax,
-                                                             uint32_t beforeStep, float afterMin, float afterMax,
+std::unique_ptr<SetNumRangeAction> SetNumRangeAction::create(const QUuid &uuid, double beforeMin, double beforeMax,
+                                                             uint32_t beforeStep, double afterMin, double afterMax,
                                                              uint32_t afterStep, AxiomModel::ModelRoot *root) {
     return std::make_unique<SetNumRangeAction>(uuid, beforeMin, beforeMax, beforeStep, afterMin, afterMax, afterStep,
                                                root);
