+++ editor/model/objects/GraphControl.h
@@ -13,9 +13,9 @@ namespace AxiomModel {
 
     struct GraphControlCurveState {
         uint8_t curveCount;
-        float curveStartVals[GRAPH_CONTROL_CURVE_COUNT + 1];
-        float curveEndPositions[GRAPH_CONTROL_CURVE_COUNT];
-        float curveTension[GRAPH_CONTROL_CURVE_COUNT];
+        double curveStartVals[GRAPH_CONTROL_CURVE_COUNT + 1];
+        double curveEndPositions[GRAPH_CONTROL_CURVE_COUNT];
+        double curveTension[GRAPH_CONTROL_CURVE_COUNT];
         uint8_t curveStates[GRAPH_CONTROL_CURVE_COUNT + 1];
     };
 
@@ -52,15 +52,15 @@ namespace AxiomModel {
 
         void setScroll(float scroll);
 
-        std::optional<uint8_t> determineInsertIndex(float time);
+        std::optional<uint8_t> determineInsertIndex(double time);
 
-        void insertPoint(uint8_t index, float time, float val, float tension, uint8_t curveState);
+        void insertPoint(uint8_t index, double time, double val, double tension, uint8_t curveState);
 
-        void movePoint(uint8_t index, float time, float value);
+        void movePoint(uint8_t index, double time, double value);
 
         void setPointTag(uint8_t index, uint8_t tag);
 
-        void setCurveTension(uint8_t index, float tension);
+        void setCurveTension(uint8_t index, double tension);
 
         void removePoint(uint8_t index);
